Fix highlight and shadow values
