+++ fuji_profiles/X-Trans-V/Fujicolor Superia 100.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.33</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>AUTO</WideDRange>
+        <DynamicRange>AUTO</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>ClassicNEGA</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>0</WBShiftR>
         <WBShiftB>-1</WBShiftB>
         <WBColorTemp>0K</WBColorTemp>
-        <HighlightTone>-1.0</HighlightTone>
-        <ShadowTone>-2.0</ShadowTone>
-        <Color>+1</Color>
-        <Sharpness>-2.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <HighlightTone>-1</HighlightTone>
+        <ShadowTone>-2</ShadowTone>
+        <Color>1</Color>
+        <Sharpness>-2</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
