+++ fuji_profiles/X-Trans-V/Ilford FP4 Plus 125.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>-1.0</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>200</WideDRange>
+        <DynamicRange>200</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>BW</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -34,8 +34,8 @@
         <HighlightTone>-0.5</HighlightTone>
         <ShadowTone>-1.5</ShadowTone>
         <Color>0</Color>
-        <Sharpness>0.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Sharpness>0</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>+2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
