+++ fuji_profiles/X-Trans-V/Kodachrome 64.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.0</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>200</WideDRange>
+        <DynamicRange>200</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>Classic</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>2</WBShiftR>
         <WBShiftB>-5</WBShiftB>
         <WBColorTemp>0K</WBColorTemp>
-        <HighlightTone>0.0</HighlightTone>
+        <HighlightTone>0</HighlightTone>
         <ShadowTone>0.5</ShadowTone>
-        <Color>+2</Color>
-        <Sharpness>1.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Color>2</Color>
+        <Sharpness>1</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>+3</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
