+++ fuji_profiles/X-Trans-V/Kodak Portra 160 v2.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.33</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>AUTO</WideDRange>
+        <DynamicRange>AUTO</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>Classic</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -34,8 +34,8 @@
         <HighlightTone>0</HighlightTone>
         <ShadowTone>0</ShadowTone>
         <Color>0</Color>
-        <Sharpness>-1.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Sharpness>-1</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
