+++ fuji_profiles/X-Trans-V/Kodak Portra 400 v2.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.33</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>400</WideDRange>
+        <DynamicRange>400</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>Classic</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>1</WBShiftR>
         <WBShiftB>-6</WBShiftB>
         <WBColorTemp>5200K</WBColorTemp>
-        <HighlightTone>0.0</HighlightTone>
-        <ShadowTone>-2.0</ShadowTone>
-        <Color>+2</Color>
-        <Sharpness>-2.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <HighlightTone>0</HighlightTone>
+        <ShadowTone>-2</ShadowTone>
+        <Color>2</Color>
+        <Sharpness>-2</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
