+++ fuji_profiles/X-Trans-V/Nostalgia Negative.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.0</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>400</WideDRange>
+        <DynamicRange>400</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>NostalgicNEGA</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>3</WBShiftR>
         <WBShiftB>-3</WBShiftB>
         <WBColorTemp>0K</WBColorTemp>
-        <HighlightTone>-1.0</HighlightTone>
-        <ShadowTone>1.0</ShadowTone>
-        <Color>+4</Color>
-        <Sharpness>-1.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <HighlightTone>-1</HighlightTone>
+        <ShadowTone>1</ShadowTone>
+        <Color>4</Color>
+        <Sharpness>-1</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-3</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
