+++ fuji_profiles/X-Trans-V/Pushed CineStill 800T.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>-0.33</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>400</WideDRange>
+        <DynamicRange>400</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>BleachBypass</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -33,9 +33,9 @@
         <WBColorTemp>7700K</WBColorTemp>
         <HighlightTone>-0.5</HighlightTone>
         <ShadowTone>1.5</ShadowTone>
-        <Color>+3</Color>
-        <Sharpness>0.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Color>3</Color>
+        <Sharpness>0</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-3</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
