+++ fuji_profiles/X-Trans-V/Summer of 1960.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.33</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>400</WideDRange>
+        <DynamicRange>400</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>NostalgicNEGA</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>-3</WBShiftR>
         <WBShiftB>-5</WBShiftB>
         <WBColorTemp>5250K</WBColorTemp>
-        <HighlightTone>4.0</HighlightTone>
-        <ShadowTone>2.0</ShadowTone>
-        <Color>+3</Color>
-        <Sharpness>-4.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <HighlightTone>4</HighlightTone>
+        <ShadowTone>2</ShadowTone>
+        <Color>3</Color>
+        <Sharpness>-4</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-3</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
