+++ fuji_profiles/X-Trans-V/Superia Xtra 400.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.0</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>400</WideDRange>
+        <DynamicRange>400</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>ClassicNEGA</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>3</WBShiftR>
         <WBShiftB>-5</WBShiftB>
         <WBColorTemp>0K</WBColorTemp>
-        <HighlightTone>0.0</HighlightTone>
-        <ShadowTone>-1.0</ShadowTone>
-        <Color>+4</Color>
-        <Sharpness>-1.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <HighlightTone>0</HighlightTone>
+        <ShadowTone>-1</ShadowTone>
+        <Color>4</Color>
+        <Sharpness>-1</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
