+++ fuji_profiles/X-Trans-V/Thommys Ektachrome.FP1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>0.0</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>100</WideDRange>
+        <DynamicRange>100</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>NostalgicNEGA</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -33,9 +33,9 @@
         <WBColorTemp>5000K</WBColorTemp>
         <HighlightTone>1.5</HighlightTone>
         <ShadowTone>1.5</ShadowTone>
-        <Color>+1</Color>
-        <Sharpness>-1.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Color>1</Color>
+        <Sharpness>-1</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
