+++ fuji_profiles/X-Trans-V/Timeless Negative.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>-0.67</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>400</WideDRange>
+        <DynamicRange>400</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>NostalgicNEGA</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>2</WBShiftR>
         <WBShiftB>-3</WBShiftB>
         <WBColorTemp>0K</WBColorTemp>
-        <HighlightTone>2.0</HighlightTone>
-        <ShadowTone>-2.0</ShadowTone>
+        <HighlightTone>2</HighlightTone>
+        <ShadowTone>-2</ShadowTone>
         <Color>-3</Color>
-        <Sharpness>0.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Sharpness>0</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
