+++ fuji_profiles/X-Trans-V/Vintage Bronze.fp1
@@ -16,8 +16,8 @@
         <ImageSize>L3x2</ImageSize>
         <ImageQuality>Normal</ImageQuality>
         <ExposureBias>-0.67</ExposureBias>
-        <DynamicRange></DynamicRange>
-        <WideDRange>200</WideDRange>
+        <DynamicRange>200</DynamicRange>
+        <WideDRange></WideDRange>
         <FilmSimulation>BleachBypass</FilmSimulation>
         <BlackImageTone></BlackImageTone>
         <MonochromaticColor_RG></MonochromaticColor_RG>
@@ -31,11 +31,11 @@
         <WBShiftR>6</WBShiftR>
         <WBShiftB>-8</WBShiftB>
         <WBColorTemp>0K</WBColorTemp>
-        <HighlightTone>0.0</HighlightTone>
-        <ShadowTone>-1.0</ShadowTone>
+        <HighlightTone>0</HighlightTone>
+        <ShadowTone>-1</ShadowTone>
         <Color>0</Color>
-        <Sharpness>-2.0</Sharpness>
-        <NoisReduction>-4.0</NoisReduction>
+        <Sharpness>-2</Sharpness>
+        <NoisReduction>-4</NoisReduction>
         <Clarity>-2</Clarity>
         <LensModulationOpt>ON</LensModulationOpt>
         <ColorSpace>sRGB</ColorSpace>
