+++ scrape/scraper.py
@@ -1,9 +1,8 @@
 import logging
 import os
 import re
-from collections.abc import Callable, Generator
+from collections.abc import Generator
 from dataclasses import dataclass
-from typing import Any, ClassVar
 
 import requests
 from bs4 import BeautifulSoup
@@ -11,45 +10,15 @@ from bs4.element import Tag
 from jinja2 import Environment, FileSystemLoader, Template
 from requests.exceptions import RequestException
 
-from scrape.models import (
-    DynamicRange,
-    FilmSimulation,
-    FujiEffect,
-    FujiSensor,
-    FujiSimulationProfile,
-    GrainEffect,
-    GrainEffectSize,
-    MonochomaticColor,
-    WhiteBalance,
-    WhiteBalanceBlueRed,
-    WhiteBalanceSetting,
-)
+from scrape.models import FilmSimulation, FujiSensor, FujiSimulationProfile, KeyStandardizer, clean_camera_profile_name
 
 # Configure logging
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 )
 
-# Example usage
 logger = logging.getLogger(__name__)
 
-
-def convert_to_float(value_str: str) -> float:
-    """
-    Converts a string to a float. Handles both regular numbers and fractions.
-    Rounds the result to two decimal places.
-
-    Args:
-    value_str (str): The string to convert, can be a number or a fraction (e.g., '1/2').
-
-    Returns:
-    float: Rounded float value of the input string.
-    """
-    if "/" in value_str:  # Fraction handling
-        numerator, denominator = map(float, value_str.split("/"))
-        return round(numerator / denominator, 2)
-    else:
-        return round(float(value_str), 2)
 
 
 def fill_xml_template(profile_dict: dict, template: str) -> str:
@@ -180,7 +149,7 @@ class FujiSimulationProfileParser:
             clean_value = KeyStandardizer.parse_key_and_standardise_value(
                 clean_key, value
             )
-            # logger.info("Parsing key '%s' with value '%s'", clean_key, clean_value)
+            logger.debug("Parsing key '%s' with value '%s'", clean_key, clean_value)
             profile_dict[clean_key] = clean_value
 
         return profile_dict
@@ -196,221 +165,6 @@ class FujiSimulationProfileParser:
             return None
         else:
             return fuji_profile
-
-
-def clean_camera_profile_name(camera_tag: str) -> str:
-    camera_profile = camera_tag.replace(" ", "_").replace(".", "").split("/")[0].upper()
-
-    alternative_camera_profile_names = {"CLASSIC_NEGATIVE": "CLASSIC_NEG"}
-
-    if camera_profile in alternative_camera_profile_names:
-        camera_profile = alternative_camera_profile_names[camera_profile]
-
-    return camera_profile
-
-
-@dataclass
-class KeyStandardizer:
-    _parsing_methods: ClassVar[dict[str, Callable[..., Any]]] = {}
-
-    @staticmethod
-    def clean_string(text_string: str) -> str:
-        """
-        Utility method to clean and standardize a string.
-        """
-        return text_string.replace(" ", "_").replace(",", "").upper()
-
-    @staticmethod
-    def parse_key_and_standardise_value(key: str, value: str) -> Any:
-        """
-        Standardizes the given key and value based on predefined parsing methods.
-
-        Args:
-            key (str): The key to be parsed.
-            value (str): The value to be standardized.
-
-        Returns:
-            Standardized value based on the key.
-        """
-        standardised_value = KeyStandardizer.clean_string(value)
-
-        # Use custom parsing method if available for the key
-        parsing_method = KeyStandardizer._parsing_methods.get(key)
-        if parsing_method:
-            return parsing_method(standardised_value)
-
-        return standardised_value
-
-    @staticmethod
-    def color_chrome_effect(value: str) -> str:
-        return FujiEffect[value].value
-
-    @staticmethod
-    def dynamic_range(value: str) -> str:
-        dynamic_range_map = {
-            "DRANGE_PRIORITY_(DRP)_AUTO": "DRAUTO",
-        }
-        if value in dynamic_range_map:
-            value = dynamic_range_map[value]
-
-        try:
-            dynamic_range_value = DynamicRange[value].value
-        except KeyError:
-            logger.warning("Could not parse dynamic range, setting to AUTO")
-            dynamic_range_value = DynamicRange.DRAUTO.value
-
-        return dynamic_range_value
-
-    @staticmethod
-    def exposure_compensation(value: str) -> float:
-        exposure_regex = r"[+-]?\d+(?:/\d+)?"
-        matches = re.findall(exposure_regex, value)
-        return convert_to_float(matches[0])
-
-    @staticmethod
-    def film_simluation(value: str) -> str:
-        camera_profile_value = clean_camera_profile_name(value)
-        return FilmSimulation[camera_profile_value].value
-
-    @staticmethod
-    def grain_effect(value: str) -> GrainEffect:
-        grain_effect_values = [item.strip() for item in value.split("_")]
-
-        try:
-            grain_effect = FujiEffect[
-                grain_effect_values[0]
-            ]  # Convert string to FujiEffect enum member
-            grain_effect_size = (
-                GrainEffectSize[grain_effect_values[1]]
-                if len(grain_effect_values) > 1
-                else None
-            )  # Convert string to GrainEffectSize enum member or None
-
-            return GrainEffect(
-                grain_effect=grain_effect, grain_effect_size=grain_effect_size
-            )
-        except (IndexError, KeyError):
-            logger.warning("Could not parse grain effect, setting to FujiEffect.OFF")
-            return GrainEffect(
-                grain_effect=FujiEffect.OFF
-            )  # Use FujiEffect.OFF directly without .value
-
-    @staticmethod
-    def white_balance(value: str) -> WhiteBalance:
-        def get_color_temperature(value: str) -> str:
-            """
-            Extracts the color temperature from the string
-            Defaults to 0K if no color temperature is found
-            """
-            temp_match = re.search(r"(\d+)K", value)
-            color_temp = int(temp_match.group(1)) if temp_match else None
-            color_temp_value = f"{color_temp}K" if color_temp else "0K"
-
-            return color_temp_value
-
-        def get_white_balance_setting(value: str) -> tuple[WhiteBalanceSetting, str]:
-            color_temp = get_color_temperature(value)
-
-            # Extract the temperature or setting
-            if "K" in value:
-                setting = WhiteBalanceSetting.TEMPERATURE
-            elif "AWB" in value:
-                setting = WhiteBalanceSetting.AUTO
-            else:
-                fluorescent_match = re.search(r"FLUORESCENT_(\d)", value)
-                if fluorescent_match:
-                    # Check if it's a fluorescent setting
-                    special_mappings = {
-                        "FLUORESCENT_1": WhiteBalanceSetting.FLIGHT1,
-                        "FLUORESCENT_2": WhiteBalanceSetting.FLIGHT2,
-                        "FLUORESCENT_3": WhiteBalanceSetting.FLIGHT3,
-                    }
-                    flight_number = fluorescent_match.group(1)
-                    setting = special_mappings[f"FLUORESCENT_{flight_number}"]
-                else:
-                    setting_match = re.match(r"([^_]+)", value)
-                    setting_name = (
-                        setting_match.group(1).upper() if setting_match else "AUTO"
-                    )
-                    setting = WhiteBalanceSetting[setting_name]
-
-            return setting, color_temp
-
-        def get_blue_red_numeric_value(
-            value: str, blue_or_red: WhiteBalanceBlueRed
-        ) -> int:
-            "Extracts the blue or red (+-) integer value from the string"
-            numeric_red_blue_regex = {
-                WhiteBalanceBlueRed.BLUE: r"([+-]?\d+)_BLUE",
-                WhiteBalanceBlueRed.RED: r"([+-]?\d+)_RED",
-            }
-
-            related_regex = numeric_red_blue_regex[blue_or_red]
-            match = re.search(related_regex, value)
-            return int(match.group(1)) if match else 0
-
-        # Extract the blue and red values / color temperature
-        blue = get_blue_red_numeric_value(value, WhiteBalanceBlueRed.BLUE)
-        red = get_blue_red_numeric_value(value, WhiteBalanceBlueRed.RED)
-        # Defaults
-        setting, color_temp = get_white_balance_setting(value)
-
-        return WhiteBalance(setting=setting, red=red, blue=blue, color_temp=color_temp)
-
-    @staticmethod
-    def numerical_values(value: str) -> float:
-        number_regex = r"([+-]?\d+(\.\d+)?)"
-        match = re.search(number_regex, value)
-        if match:
-            converted_value = float(match.group(0))
-        else:
-            logger.warning("Could not convert %s to float, setting to 0", value)
-            converted_value = 0.0
-
-        return converted_value
-
-    @staticmethod
-    def monochromatic_color(value: str) -> MonochomaticColor:
-        """
-        Extracts the monochromatic color from the string
-        Example:
-            value: Monochromatic Color: 0 WC & 0 MG
-        Return:
-            0
-        """
-        monochromatic_color_regex = r"(\d+)_WC_&_(\d+)_MG"
-        match = re.search(monochromatic_color_regex, value)
-
-        if match:
-            warm_cool = int(match.group(1))
-            magenta_green = int(match.group(2))
-        else:
-            logger.warning(
-                "Could not convert %s to MonochromaticColor, setting to 0", value
-            )
-            warm_cool = 0
-            magenta_green = 0
-
-        return MonochomaticColor(
-            warm_cool=warm_cool,
-            magenta_green=magenta_green,
-        )
-
-    @classmethod
-    def initialise_parsing_methods(cls) -> None:
-        cls._parsing_methods = {
-            "color_chrome_effect": KeyStandardizer.color_chrome_effect,
-            "dynamic_range": KeyStandardizer.dynamic_range,
-            "exposure_compensation": KeyStandardizer.exposure_compensation,
-            "film_simulation": KeyStandardizer.film_simluation,
-            "grain_effect": KeyStandardizer.grain_effect,
-            "high_iso_nr": KeyStandardizer.numerical_values,
-            "highlight": KeyStandardizer.numerical_values,
-            "shadow": KeyStandardizer.numerical_values,
-            "sharpness": KeyStandardizer.numerical_values,
-            "white_balance": KeyStandardizer.white_balance,
-            "monochromatic_color": KeyStandardizer.monochromatic_color,
-        }
 
 
 @dataclass
@@ -485,7 +239,7 @@ class FujiRecipe:
     def filled_template(self) -> str:
         "Returns a filled Jinja2 template as a string"
         initial_filled_template = self.jinja2_template.render(self.link.__dict__)
-        filled_template = fill_xml_template(self.as_dict(), initial_filled_template)
+        filled_template = fill_xml_template(self.as_dict(), initial_filled_template) + "\n"
         return filled_template
 
     def as_dict(self) -> dict:
@@ -591,4 +345,5 @@ if __name__ == "__main__":
         related_recipes = FujiRecipes.fetch_recipes(sensor, sensor_url)
 
         for recipe in related_recipes:
+            # if recipe.link.name == "1976 Kodak":
                 recipe.save()
